Moving Search Results for its component
